Moving test locations to remove environment information.
