+++ tests/notifications/functional_tests/test_email_auth.py
