+++ tests/notifications/functional_tests/test_join_live_service.py
