+++ tests/notifications/functional_tests/test_registration_and_invite.py
