+++ tests/notifications/smoke_tests/notify_api/test_notify_api_email.py
