+++ tests/notifications/smoke_tests/test_admin.py
